move status components to status.rs
