+++ rocksdb-cxx/include/rocksdb/io_status.h
@@ -24,9 +24,9 @@
 #include "status.h"
 
 #ifndef ROCKSDB_RS
-#include "rocksdb-rs-cxx/lib.h"
+#include "rocksdb-rs-cxx/status.h"
 #else
-#include "rocksdb-rs/src/lib.rs.h"
+#include "rocksdb-rs/src/status.rs.h"
 #endif
 
 namespace ROCKSDB_NAMESPACE {
